apps/api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pyjwt[crypto]==2.10.1
